requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic>=2.8.0,<3
python-dotenv==1.0.1
loguru==0.7.2
PyYAML==6.0.2
# 轻量语义检索（无需 torch/编译）
fastembed==0.7.1
# 分词与文本处理（可选）
jieba==0.42.1
rapidfuzz==3.9.3
# 数据处理工具
pandas>=2.0.0
openpyxl>=3.1.0
# 测试与客户端
pytest==8.2.1
httpx==0.27.0
requests==2.32.4

# Flask Web框架
Flask==2.3.3
Werkzeug==2.3.7
Jinja2==3.1.2

# 备选语义检索（可选，未安装时自动降级）
# faiss-cpu==1.8.0.post1
# transformers==4.41.2
# torch==2.3.1+cpu
# sentence-transformers==2.7.0
